refactor: direct to new functions
