+++ io/limited_reader_test.ts
@@ -3,7 +3,7 @@ import { assertEquals } from "../assert/mod.ts";
 import { LimitedReader } from "./limited_reader.ts";
 import { StringWriter } from "./string_writer.ts";
 import { copy } from "../streams/copy.ts";
-import { readAll } from "../streams/read_all.ts";
+import { readAll } from "../io/read_all.ts";
 import { StringReader } from "./string_reader.ts";
 
 Deno.test("ioLimitedReader", async function () {
